Turn vsync into a core service.
Avoids loading it redundantly into extensions.

Part of #1055
