+++ src/amp-core-service.js
@@ -17,6 +17,7 @@
 import {installHistoryService} from './service/history-impl';
 import {installViewerService} from './service/viewer-impl';
 import {installViewportService} from './service/viewport-impl';
+import {installVsyncService} from './service/vsync-impl';
 
 
 /**
@@ -28,4 +29,5 @@ export function installCoreServices(window) {
   installViewerService(window);
   installViewportService(window);
   installHistoryService(window);
+  installVsyncService(window);
 }
